refactor: 상수 경로 수정
