+++ src/components/registerSection/registerCardIssuer.style.ts
@@ -1,0 +1,6 @@
+import { styled } from 'styled-components';
+
+export const Select = styled.select`
+  width: 100%;
+  padding: 8px;
+`;
