+++ src/constants/cardSection.ts
@@ -5,8 +5,16 @@ export const CARD_NUMBER = Object.freeze({
   description: '본인 명의의 카드만 결제 가능합니다.',
   inputTitle: '카드 번호',
   errorMessage: NUMBER_ERROR_MESSAGE,
-  INDIVIDUAL_MAX_LENGTH: 4,
-  TOTAL_MAX_LENGTH: 16,
+});
+
+export const MAX_LENGTH = Object.freeze({
+  TOTAL_CARD_NUMBER: 16,
+  INDIVIDUAL_CARD_NUMBER: 4,
+  MONTH: 2,
+  YEAR: 2,
+  NAME: 30,
+  CVC: 3,
+  PASSWORD: 2,
 });
 
 export const EXPIRATION_PERIOD = Object.freeze({
@@ -15,8 +23,6 @@ export const EXPIRATION_PERIOD = Object.freeze({
   inputTitle: '유효기간',
   monthErrorMessage: '1부터 12사이의 숫자만 입력 가능합니다.',
   yearErrorMessage: NUMBER_ERROR_MESSAGE,
-  MONTH_MAX_LENGTH: 2,
-  YEAR_MAX_LENGTH: 2,
 });
 
 export const OWNER_NAME = Object.freeze({
@@ -34,7 +40,6 @@ export const CARD_CVC = Object.freeze({
   title: 'CVC 번호를 입력해 주세요.',
   inputTitle: 'CVC',
   errorMessage: '숫자 3개를 입력해 주세요',
-  MAX_LENGTH: 3,
 });
 
 export const PASSWORD = Object.freeze({
@@ -42,5 +47,4 @@ export const PASSWORD = Object.freeze({
   description: '앞의 2자리를 입력해 주세요',
   inputTitle: '비밀번호 앞 2자리',
   errorMessage: '숫자만 입력해 주세요',
-  MAX_LENGTH: 2,
 });
